refactor(memory): split hook into per-tool files, fix "memory" tool routing
The "memory" tool registered as `memory_tool` (snake_case of the
`MemoryTool` struct name) but `dispatch()` matched `"memory"`, so the
tool never actually routed. Pin the wire name with
`#[schemars(rename = "Memory")]` on the renamed `Prompt` struct.

Mirror the os/ hook layout: each tool gets its own file owning the
input struct, the `Memory::method`, and the `MemoryHook::handle_X`
dispatch handler. The `Hook` impl stays in `mod.rs` since it isn't a
tool. Add `store_read`/`store_write` helpers so the `RwLock` unwrap
contract lives in one place.

Drop the legacy `memory/` directory migration — this is a feature
branch, no released code depends on it. Update the prompt that still
described `remember(name, description, content)` instead of the actual
`(name, content, aliases)` signature.
