+++ crates/crabtalk/tests/memory.rs
@@ -1,13 +1,12 @@
 //! Integration tests for the hook-level memory facade.
 
 use crabtalk::hooks::Memory;
-use std::fs;
 use tempfile::tempdir;
 use wcore::MemoryConfig;
 
 fn test_memory() -> Memory {
     let dir = tempdir().unwrap();
-    Memory::open(MemoryConfig::default(), dir.path().join("memory.db"), None).unwrap()
+    Memory::open(MemoryConfig::default(), dir.path().join("memory.db")).unwrap()
 }
 
 #[test]
@@ -160,44 +159,6 @@ fn aliases_boost_search() {
 }
 
 #[test]
-fn migration_imports_legacy_entries_and_index() {
-    // Arrange a legacy on-disk layout: config_dir/memory/entries/*.md
-    // + config_dir/memory/MEMORY.md.
-    let dir = tempdir().unwrap();
-    let config_dir = dir.path();
-    let legacy = config_dir.join("memory");
-    let entries = legacy.join("entries");
-    fs::create_dir_all(&entries).unwrap();
-
-    fs::write(
-        entries.join("luna.md"),
-        "---\nname: luna\ndescription: User's dog Luna is a golden retriever\n---\n\nLuna has vet visits on Thursdays.\n",
-    )
-    .unwrap();
-    fs::write(
-        entries.join("rust.md"),
-        "---\nname: rust\ndescription: User works on crabtalk\n---\n\nCrabtalk is an AI companion.\n",
-    )
-    .unwrap();
-    fs::write(legacy.join("MEMORY.md"), "# My overview\n\nSome prose.").unwrap();
-
-    let mem = Memory::open(
-        MemoryConfig::default(),
-        config_dir.join("memory.db"),
-        Some(legacy),
-    )
-    .unwrap();
-
-    let golden = mem.recall("golden retriever", 5);
-    assert!(golden.contains("Luna"));
-    let rust = mem.recall("crabtalk", 5);
-    assert!(rust.contains("Crabtalk"));
-
-    let prompt = mem.build_prompt();
-    assert!(prompt.contains("My overview"));
-}
-
-#[test]
 fn remember_rejects_reserved_global_name() {
     let mem = test_memory();
     let result = mem.remember("global".into(), "hijacked".into(), vec![]);
@@ -214,82 +175,4 @@ fn forget_rejects_reserved_global_name() {
     assert!(result.contains("reserved"), "got: {result}");
     // Prompt content survives.
     assert!(mem.build_prompt().contains("overview"));
-}
-
-#[test]
-fn migration_creates_db_even_when_all_legacy_entries_malformed() {
-    let dir = tempdir().unwrap();
-    let config_dir = dir.path();
-    let legacy = config_dir.join("memory");
-    let entries = legacy.join("entries");
-    fs::create_dir_all(&entries).unwrap();
-
-    // All legacy entries lack frontmatter — every `Op::Add` will be
-    // skipped, so without the unconditional checkpoint the db file
-    // would never be created and the next open would re-enter
-    // migration.
-    fs::write(entries.join("broken1.md"), "no frontmatter here").unwrap();
-    fs::write(entries.join("broken2.md"), "plain text").unwrap();
-
-    let db = config_dir.join("memory.db");
-    {
-        let _ = Memory::open(MemoryConfig::default(), db.clone(), Some(legacy.clone())).unwrap();
-    }
-    assert!(db.exists(), "db file must exist after migration attempt");
-
-    // Even a later-added valid legacy entry should NOT be imported —
-    // migration has already happened.
-    fs::write(
-        entries.join("valid.md"),
-        "---\nname: valid\ndescription: new\n---\n\nbody\n",
-    )
-    .unwrap();
-    let mem = Memory::open(MemoryConfig::default(), db, Some(legacy)).unwrap();
-    assert_eq!(mem.recall("new", 5), "no memories found");
-}
-
-#[test]
-fn migration_skipped_when_db_already_exists() {
-    let dir = tempdir().unwrap();
-    let config_dir = dir.path();
-    let legacy = config_dir.join("memory");
-    let entries = legacy.join("entries");
-    fs::create_dir_all(&entries).unwrap();
-    fs::write(
-        entries.join("existing.md"),
-        "---\nname: existing\ndescription: should not be imported\n---\n\nlegacy content\n",
-    )
-    .unwrap();
-
-    // First open creates an empty db file.
-    {
-        let _ = Memory::open(
-            MemoryConfig::default(),
-            config_dir.join("memory.db"),
-            Some(legacy.clone()),
-        )
-        .unwrap();
-    }
-
-    // Re-opening with legacy set must NOT re-import — the file already exists.
-    // To verify: the migration happens once. Mutate legacy to add a new entry,
-    // reopen, and confirm the new entry is absent.
-    fs::write(
-        entries.join("late.md"),
-        "---\nname: late\ndescription: added after first open\n---\n\nshould be skipped\n",
-    )
-    .unwrap();
-
-    let mem = Memory::open(
-        MemoryConfig::default(),
-        config_dir.join("memory.db"),
-        Some(legacy),
-    )
-    .unwrap();
-
-    let result = mem.recall("skipped", 5);
-    assert_eq!(
-        result, "no memories found",
-        "migration should not re-run on an existing db"
-    );
 }
